+++ packages/yourloops/pages/patient-view/patient-profile/patient-profile-view.tsx
@@ -28,7 +28,7 @@
 import React, { type FC, useState } from 'react'
 import Container from '@mui/material/Container'
 import Grid from '@mui/material/Grid'
-import { PatientPersonalInformationSection } from './sections/patient-personal-information-section'
+import { PatientPersonalInformationSection } from './sections/personal-information/patient-personal-information-section'
 import { PatientProfileViewMenu } from './patient-profile-view-menu'
 import { PatientProfileViewSection } from './patient-profile-view-section.enum'
 import { AlertsSection } from './sections/alerts-section'