+++ test/locales/translations.test.js
@@ -331,7 +331,10 @@ const ignoredTransKeyForYourLoops = [
   'device-change-pump-name',
   'device-change-mobile-app-activation-code',
   'device-change-sw-version',
-  'device-history-title'
+  'device-history-title',
+  'add-clinician-description',
+  'remove-clinician-question',
+  'remove-clinician-question-patient'
 ]
 const ignoredTransKeyInYourLoopsFiles = [
   // Documentation!